Update chapter/lab expectations_b200.json after strict revalidation. Fix correctness/stability in ch06/ch09/ch10/ch15 and speculative/moe/ddp labs. Harden benchmark harness/tooling behavior (bench_commands, validity checks, extension loader). Update enforcement test coverage.

+++ code/ch15/baseline_speculative_decoding.py
@@ -23,7 +23,8 @@ class BaselineSpeculativeDecodingBenchmark(VerificationPayloadMixin, BaseBenchma
     def __init__(self) -> None:
         super().__init__()
 
-        self.workload = default_workload(dtype=torch.bfloat16)
+        # Use FP32 for deterministic argmax stability across kernel launch shapes.
+        self.workload = default_workload(dtype=torch.float32)
 
         self.target_model: Optional[TokenMLP] = None
         self.input_ids: Optional[torch.Tensor] = None
