+++ code/labs/speculative_decode/baseline_speculative_decode.py
@@ -24,7 +24,8 @@ class BaselineSpeculativeDecodeBenchmark(VerificationPayloadMixin, BaseBenchmark
     def __init__(self) -> None:
         super().__init__()
 
-        self.workload = default_workload(dtype=torch.bfloat16)
+        # Use FP32 for deterministic argmax stability across launch/shape variations.
+        self.workload = default_workload(dtype=torch.float32)
 
         self.target_model: Optional[TokenMLP] = None
         self.input_ids: Optional[torch.Tensor] = None
